Updated Chrome to version 37 and added a check to see if chrome is already install since chrome will update itself...

+++ chrome/tools/chocolateyInstall.ps1
@@ -8,9 +8,15 @@ if ($psISE) {
 }
 
 try {
-    Install-ChocolateyPackage $packageName $installerType $installerArgs $url $url64
+    $app = Get-WmiObject -Class Win32_Product | Where-Object { $_.Name -eq "Google Chrome" }
+    
+    if ($app) {
+        Write-Output "Chrome is already installed and will update itself, no need to update via package..."
+    } else {
+        Install-ChocolateyPackage $packageName $installerType $installerArgs $url $url64
 
-    Start-ChocolateyProcessAsAdmin "Remove-Item '$($env:PUBLIC)\Desktop\Google Chrome.lnk' -Force"
+        Start-ChocolateyProcessAsAdmin "Remove-Item '$($env:PUBLIC)\Desktop\Google Chrome.lnk' -Force"
+    }
 
     Write-ChocolateySuccess $packageName
 } catch {
